feat: move CompareDrawer to ui components
