+++ .changeset/weak-humans-heal.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": patch
+---
+
+Move CompareDrawer to ui components.
