+++ core/components/ui/compare-drawer/index.ts
@@ -1,0 +1,3 @@
+'use client';
+
+export * from './context';
